Don't double-embed a NullString
The implicit embed here was supposed to put the methods from sql.NullString
on our copy of NullString but in practice sql.NullString only has two methods
(Scan() and Value()), each of which already exist on this object. The effect
is to attach one additional property to a NullString called "NullString" which
*also* has Valid and String fields, so you can do s.NullString.Valid and
s.Valid, and potentially get two different results. Needless to say this is
a bit silly and unnecessary.